fix: decouple lifecycle roster callbacks
