+++ modules/ai_team/1module.go
@@ -6,6 +6,7 @@ import (
 	"github.com/Mininglamp-OSS/octo-lib/config"
 	"github.com/Mininglamp-OSS/octo-lib/pkg/register"
 	"github.com/Mininglamp-OSS/octo-server/modules/botfather"
+	aiteampkg "github.com/Mininglamp-OSS/octo-server/pkg/aiteam"
 )
 
 //go:embed sql
@@ -13,6 +14,7 @@ var sqlFS embed.FS
 
 func init() {
 	botfather.RegisterAITeamProvisioner(ProvisionOwnedBot)
+	aiteampkg.RegisterLifecycleRosterMarker(markLifecycleRosterRemovalTx)
 	register.AddModule(func(ctx interface{}) register.Module {
 		return register.Module{
 			Name: "ai_team",
