+++ modules/group/lifecycle_group_lookup_test.go
@@ -76,47 +76,3 @@ func TestRemoveUserFromGroupsForLifecycleCleanupSkipsGroupCreator(t *testing.T) 
 	require.Equal(t, MemberRoleCreator, member.Role)
 	require.Zero(t, member.IsDeleted, "creator membership is intentionally retained without failing teardown")
 }
-
-func TestLifecycleCleanupDeactivatesAgentAndAdvancesTeamRoster(t *testing.T) {
-	svc, _, ctx := setupServiceTestWithCtx(t)
-	const (
-		spaceID  = "ai_lifecycle_space"
-		ownerUID = "ai_lifecycle_owner"
-		botUID   = "ai_lifecycle_bot"
-		groupNo  = "ai_lifecycle_team"
-	)
-
-	require.NoError(t, NewDB(ctx).Insert(&Model{
-		GroupNo: groupNo, Name: aiteam.TeamGroupName, Creator: ownerUID,
-		SpaceID: spaceID, Status: GroupStatusNormal, Purpose: aiteam.TeamGroupPurpose,
-	}))
-	for _, member := range []*MemberModel{
-		{GroupNo: groupNo, UID: ownerUID, Role: MemberRoleCreator, Status: int(common.GroupMemberStatusNormal), Version: 1},
-		{GroupNo: groupNo, UID: botUID, Role: MemberRoleCommon, Status: int(common.GroupMemberStatusNormal), Robot: 1, Version: 2},
-	} {
-		require.NoError(t, NewDB(ctx).InsertMember(member))
-	}
-	_, err := ctx.DB().InsertBySql(`INSERT INTO ai_team_agent
-		(space_id,user_uid,bot_id,is_added,container_state) VALUES (?,?,?,?,?)`,
-		spaceID, ownerUID, botUID, 1, 2).Exec()
-	require.NoError(t, err)
-	_, err = ctx.DB().InsertBySql(`INSERT INTO ai_team_group
-		(space_id,user_uid,group_no,state,roster_version) VALUES (?,?,?,?,?)`,
-		spaceID, ownerUID, groupNo, 2, 7).Exec()
-	require.NoError(t, err)
-
-	require.NoError(t, svc.RemoveUserFromGroupsForLifecycleCleanup(botUID))
-
-	var agentAdded int
-	require.NoError(t, ctx.DB().Select("is_added").From("ai_team_agent").
-		Where("space_id=? AND user_uid=? AND bot_id=?", spaceID, ownerUID, botUID).LoadOne(&agentAdded))
-	require.Zero(t, agentAdded)
-	var projection struct {
-		State   int    `db:"state"`
-		Version uint64 `db:"roster_version"`
-	}
-	require.NoError(t, ctx.DB().Select("state", "roster_version").From("ai_team_group").
-		Where("space_id=? AND user_uid=?", spaceID, ownerUID).LoadOne(&projection))
-	require.Equal(t, aiTeamProjectionPending, projection.State)
-	require.EqualValues(t, 8, projection.Version)
-}
